Create host-transfer backend, move methods from MPI-CUDA to host-transfer.
