+++ benchmark/benchmark_nballreduces.cpp
@@ -34,6 +34,9 @@
 #ifdef AL_HAS_MPI_CUDA
 #include "test_utils_mpi_cuda.hpp"
 #endif
+#ifdef AL_HAS_HOST_TRANSFER
+#include "test_utils_ht_cuda.hpp"
+#endif
 
 size_t start_size = 1;
 size_t max_size = 1<<30;
@@ -63,6 +66,13 @@ typename Al::NCCLBackend::comm_type get_comm<Al::NCCLBackend>() {
 template <>
 typename Al::MPICUDABackend::comm_type get_comm<Al::MPICUDABackend>() {
   return typename Al::MPICUDABackend::comm_type(MPI_COMM_WORLD, bm_stream);
+}
+#endif
+
+#ifdef AL_HAS_HOST_TRANSFER
+template <>
+typename Al::HTBackend::comm_type get_comm<Al::HTBackend>() {
+  return typename Al::HTBackend::comm_type(MPI_COMM_WORLD, bm_stream);
 }
 #endif
 
@@ -260,6 +270,10 @@ int main(int argc, char** argv) {
   } else if (backend == "MPI-CUDA") {
     do_benchmark<Al::MPICUDABackend>();
 #endif    
+#ifdef AL_HAS_HOST_TRANSFER
+  } else if (backend == "HT") {
+    do_benchmark<Al::HTBackend>();
+#endif
   } else {
     std::cerr << "usage: " << argv[0] << " [MPI";
 #ifdef AL_HAS_NCCL
@@ -267,6 +281,9 @@ int main(int argc, char** argv) {
 #endif
 #ifdef AL_HAS_MPI_CUDA
     std::cerr << " | MPI-CUDA";
+#endif
+#ifdef AL_HAS_HOST_TRANSFER
+    std::cerr << " | HT";
 #endif
     std::cerr << "]" << std::endl;
     return -1;
