+++ cmake/Al_config.hpp.in
@@ -3,6 +3,7 @@
 #cmakedefine AL_HAS_CUDA
 #cmakedefine AL_HAS_MPI_CUDA
 #cmakedefine AL_HAS_NCCL
+#cmakedefine AL_HAS_HOST_TRANSFER
 
 #ifdef AL_HAS_MPI_CUDA
 #cmakedefine AL_HAS_MPI_CUDA_RMA
