+++ src/CMakeLists.txt
@@ -36,6 +36,12 @@ if (AL_HAS_NCCL)
   list(APPEND THIS_DIR_CXX_SOURCES "${CMAKE_CURRENT_SOURCE_DIR}/nccl_impl.cpp")
 endif ()
 
+if (AL_HAS_HOST_TRANSFER)
+  list(APPEND THIS_DIR_HEADERS "${CMAKE_CURRENT_SOURCE_DIR}/ht_impl.hpp")
+  list(APPEND THIS_DIR_CXX_SOURCES "${CMAKE_CURRENT_SOURCE_DIR}/ht_impl.cpp")
+  add_subdirectory(ht)
+endif()
+
 set(ALUMINUM_HEADERS "${HEADERS}" "${THIS_DIR_HEADERS}")
 set(ALUMINUM_CXX_SOURCES "${CXX_SOURCES}" "${THIS_DIR_CXX_SOURCES}")
 set(ALUMINUM_CUDA_SOURCES "${CUDA_SOURCES}" "${THIS_DIR_CUDA_SOURCES}")
