+++ src/ht/reduce_scatter.hpp
@@ -28,80 +28,66 @@
 #pragma once
 
 #include "cuda.hpp"
-#include "cuda_kernels.hpp"
-#include "mpi_cuda/communicator.hpp"
 #include "progress.hpp"
 
 namespace Al {
 namespace internal {
-namespace mpi_cuda {
+namespace host_transfer {
 
 template <typename T>
-class BcastAlState : public AlState {
+class ReduceScatterAlState : public AlState {
 public:
-  BcastAlState(T* buf, size_t count, int root,
-               MPICUDACommunicator& comm, cudaStream_t stream) :
+  ReduceScatterAlState(const T* sendbuf, T* recvbuf, size_t count,
+                       ReductionOperator op, HTCommunicator& comm,
+                       cudaStream_t stream) :
     AlState(nullptr),
-    rank_(comm.rank()), root_(root), count_(count),
-    host_mem_(get_pinned_memory<T>(count_)),
+    count_(count),
+    host_mem_(get_pinned_memory<T>(comm.size()*count_)),
+    op_(mpi::ReductionOperator2MPI_Op(op)),
     comm_(comm.get_comm()),
     compute_stream(comm.get_stream()) {
 
-    bool const i_am_root = rank_ == root_;
-
     // Transfer data from device to host and use an event to determine when it
     // completes.
-    if (i_am_root) {
-      AL_CHECK_CUDA(cudaMemcpyAsync(
-                      host_mem_, buf, sizeof(T)*count_,
-                      cudaMemcpyDeviceToHost, stream));
-    }
+    AL_CHECK_CUDA(cudaMemcpyAsync(host_mem_, sendbuf, sizeof(T)*count*comm.size(),
+                                  cudaMemcpyDeviceToHost, stream));
     d2h_event_.record(stream);
+
+    // Have the device wait on the host.
     gpuwait_.wait(stream);
 
-    if (!i_am_root) {
-      // Transfer completed buffer back to device.
-      AL_CHECK_CUDA(cudaMemcpyAsync(buf, host_mem_, sizeof(T)*count_,
-                                    cudaMemcpyHostToDevice, stream));
-      h2d_event_.record(stream);
-    }
+    // Transfer completed buffer back to device.
+    AL_CHECK_CUDA(cudaMemcpyAsync(recvbuf, host_mem_, sizeof(T)*count,
+                                  cudaMemcpyHostToDevice, stream));
+    h2d_event_.record(stream);
   }
 
-  ~BcastAlState() override {
+  ~ReduceScatterAlState() override {
     release_pinned_memory(host_mem_);
   }
 
   bool step() override {
-    if (!bcast_started_) {
+    if (!rs_started_) {
+      // Wait for memory to get to the host.
       if (d2h_event_.query()) {
-        MPI_Ibcast(host_mem_, count_, mpi::TypeMap<T>(),
-                     root_, comm_, &req_);
-        if (rank_ == root_) {
-          gpuwait_.signal();
-        }
-        bcast_started_ = true;
+        MPI_Ireduce_scatter_block(MPI_IN_PLACE, host_mem_, count_,
+                                  mpi::TypeMap<T>(), op_, comm_, &req_);
+        rs_started_ = true;
       } else {
         return false;
       }
     }
-
-    if (!bcast_done_) {
-      // Wait for the bcast to complete
+    if (!rs_done_) {
+      // Wait for the RS to complete.
       int flag;
       MPI_Test(&req_, &flag, MPI_STATUS_IGNORE);
       if (flag) {
-        bcast_done_ = true;
-        if (rank_ != root_) {
-          gpuwait_.signal();
-        } else {
-          return true;
-        }
-      }
-      else {
+        rs_done_ = true;
+        gpuwait_.signal();
+      } else {
         return false;
       }
     }
-
     // Wait for host-to-device memcopy; cleanup
     if (h2d_event_.query()) {
       return true;
@@ -113,24 +99,18 @@ public:
   void* get_compute_stream() const override { return compute_stream; }
 
 private:
-  int rank_;
-  int root_;
   size_t count_;
   T* host_mem_;
-
   cuda::GPUWait gpuwait_;
-
   cuda::FastEvent d2h_event_, h2d_event_;
-
+  MPI_Op op_;
   MPI_Comm comm_;
   MPI_Request req_ = MPI_REQUEST_NULL;
-
-  bool bcast_started_ = false;
-  bool bcast_done_ = false;
-
+  bool rs_started_ = false;
+  bool rs_done_ = false;
   cudaStream_t compute_stream;
 };
 
-}  // namespace mpi_cuda
+}  // namespace host_transfer
 }  // namespace internal
 }  // namespace Al
