+++ src/mpi_cuda/CMakeLists.txt
@@ -1,17 +1,9 @@
 set_full_path(THIS_DIR_HEADERS
-  allgather.hpp
   allreduce.hpp
   allreduce_ring.hpp
-  alltoall.hpp
-  bcast.hpp
   communicator.hpp
   cuda_kernels.hpp
   cuda_local_transfer.hpp
-  gather.hpp
-  reduce.hpp
-  reduce_scatter.hpp
-  scatter.hpp
-  pt2pt.hpp
   util.hpp
   )
