+++ test/test_allreduce.cpp
@@ -34,6 +34,9 @@
 #ifdef AL_HAS_MPI_CUDA
 #include "test_utils_mpi_cuda.hpp"
 #endif
+#ifdef AL_HAS_HOST_TRANSFER
+#include "test_utils_ht_cuda.hpp"
+#endif
 
 #include <stdlib.h>
 #include <math.h>
@@ -156,6 +159,10 @@ int main(int argc, char** argv) {
 #ifdef AL_HAS_MPI_CUDA
   } else if (backend == "MPI-CUDA") {
     test_correctness<Al::MPICUDABackend>();
+#endif
+#ifdef AL_HAS_HOST_TRANSFER
+  } else if (backend == "HT") {
+    test_correctness<Al::HTBackend>();
 #endif
   }
